refactor: better implementation using component
