+++ components/Header.jsx
@@ -1,0 +1,11 @@
+import "../styles.css";
+
+function Header() {
+  return (
+    <header className="header">
+      <h1>Martabak Mang Ujang</h1>
+    </header>
+  );
+}
+
+export default Header;
